First draft of boilerplate template. Not yet working.
This commit contains the basic structure and files, but Javascript logic needs to be re-invented, because the Elastic Caller should be called by a JS REST call once the user clicks "search". Therefore the Elastic Caller has to be static.

Beyond that, API keys for Google and OpenCage shall be moved to wp-config.

Right now the orchestration of the two searches - first one for entered location in OpenCage, second one for clubs in range by the Elastic Caller - is implemented in JS. Those two calls should be combined on server side.